src/requirements.txt
psycopg2-binary
pandas==1.4.2
boto3==1.24.10
python-dateutil==2.8.2
requests==2.28.0